Factor other files to separate component
